fix (server): wrong variables getter
