--- server/src/flows/helpers/get-variable.ts
@@ -1,9 +1,0 @@
-export const getVariable = (
-  arg: string,
-  variables: Array<{
-    reference: string;
-    value: unknown;
-  }>,
-): unknown | undefined => {
-  return variables.reverse().find(({ reference }) => reference === arg)?.value;
-};
