+++ server/src/flows/helpers/index.ts
@@ -1,3 +1,3 @@
 export * from './get-menu';
 export * from './context-builder';
-export * from './get-variable';
+export * from './variables';
